requirements.txt
aioredis==1.3.1
asgi-redis==1.0.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
boto3==1.18.63
botocore==1.21.63
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.6
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-allauth==0.45.0
django-filter==21.1
django-import-export==2.6.1
django-storages==1.12.2
et-xmlfile==1.1.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
jmespath==0.10.0
MarkupPy==1.14
msgpack==1.0.2
msgpack-python==0.5.6
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
